Issue #776 Move org.ehcache.core.util to internal
* Moved most types to org.ehcache.core.internal.util
* Deleted KeysIterable and LifeCycleUtils
