+++ core/src/main/java/org/ehcache/core/EhcacheWithLoaderWriter.java
@@ -71,7 +71,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 
-import static org.ehcache.core.util.Functions.memoize;
+import static org.ehcache.core.internal.util.Functions.memoize;
 import static org.ehcache.core.exceptions.ExceptionFactory.newCacheLoadingException;
 import static org.ehcache.core.exceptions.ExceptionFactory.newCacheWritingException;
 import static org.terracotta.statistics.StatisticBuilder.operation;
