+++ impl/src/main/java/org/ehcache/config/builders/UserManagedCacheBuilder.java
@@ -60,7 +60,7 @@ import org.ehcache.spi.service.Service;
 import org.ehcache.spi.service.ServiceConfiguration;
 import org.ehcache.spi.service.ServiceCreationConfiguration;
 import org.ehcache.spi.service.ServiceDependencies;
-import org.ehcache.core.util.ClassLoading;
+import org.ehcache.core.internal.util.ClassLoading;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
