+++ xml/src/main/java/org/ehcache/xml/ConfigurationParser.java
@@ -43,7 +43,7 @@ import org.ehcache.xml.model.SizeofType;
 import org.ehcache.xml.model.TimeType;
 import org.ehcache.spi.service.ServiceConfiguration;
 import org.ehcache.spi.service.ServiceCreationConfiguration;
-import org.ehcache.core.util.ClassLoading;
+import org.ehcache.core.internal.util.ClassLoading;
 import org.w3c.dom.Element;
 import org.xml.sax.ErrorHandler;
 import org.xml.sax.SAXException;
